+++ compiler/src/lexer.rs
@@ -17,16 +17,20 @@ pub fn lex(source_code: &str) -> Result<Vec<RichToken>, LexingError> {
     let mut tokens = vec![];
 
     while let Some(token) = l.next() {
-        let token = token.map_err(|e| match e {
+        let kind = token.map_err(|e| match e {
             LexingError::Default => LexingError::UnknownToken {
                 token: String::from(l.slice()),
                 span: l.span(),
             },
             _ => e,
-        });
+        })?;
+
+        if kind.is_comment() {
+            continue;
+        }
 
         tokens.push(RichToken {
-            kind: token?,
+            kind,
             span: l.span(),
             value: String::from(l.slice()),
         })
@@ -80,9 +84,12 @@ pub enum Token {
     // Keywords
     //
 
-    #[token("var")]  Var,
-    #[token("if")]   If,
-    #[token("else")] Else,
+    #[token("var")]      Var,
+    #[token("if")]       If,
+    #[token("else")]     Else,
+    #[token("while")]    While,
+    #[token("break")]    Break,
+    #[token("continue")] Continue,
 
     //
     // Symbols
@@ -92,6 +99,7 @@ pub enum Token {
     #[token("-")] Sub,
     #[token("*")] Mul,
     #[token("/")] Div,
+    #[token("%")] Mod,
 
     #[token(":")] Colon,
     #[token(";")] Semicolon,
@@ -109,9 +117,27 @@ pub enum Token {
     #[token("<")]  Lt,
     #[token("<=")] Lte,
 
+    #[token("||")] Or,
+    #[token("&&")] And,
+    #[token("!")]  Not,
+
+    // Comments
+
+    #[token("//", callback = single_line_comment)]
+    SingleLineComment(String),
+
+    #[token("/*", callback = multi_line_comment)]
+    MultiLineComment(String),
+
     // This will always be appended as the last token
     // inside token list
     Eof,
+}
+
+impl Token {
+    fn is_comment(&self) -> bool {
+        matches!(self, Token::SingleLineComment(_)) || matches!(self, Token::MultiLineComment(_))
+    }
 }
 
 impl Display for Token {
@@ -125,10 +151,14 @@ impl Display for Token {
             Self::Var => write!(f, "`var` keyword"),
             Self::If => write!(f, "`if` keyword"),
             Self::Else => write!(f, "`else` keyword"),
+            Self::While => write!(f, "`while` keyword"),
+            Self::Break => write!(f, "`break` keyword"),
+            Self::Continue => write!(f, "`continue` keyword"),
             Self::Add => write!(f, "addition operator (`+`)"),
             Self::Sub => write!(f, "subtraction operator (`-`)"),
             Self::Mul => write!(f, "multiplication operator (`*`)"),
             Self::Div => write!(f, "division operator (`/`)"),
+            Self::Mod => write!(f, "modulo operator (`%`)"),
             Self::Colon => write!(f, "colon (`:`)"),
             Self::Semicolon => write!(f, "semicolon (`;`)"),
             Self::Comma => write!(f, "comma (`,`)"),
@@ -143,6 +173,12 @@ impl Display for Token {
             Self::Gte => write!(f, "greater than or equal operator (`>=`)"),
             Self::Lt => write!(f, "less than operator (`<`)"),
             Self::Lte => write!(f, "less than or equal operator (`<=`)"),
+            Self::Or => write!(f, "logical or operator (`||`)"),
+            Self::And => write!(f, "logical and operator (`&&`)"),
+            Self::Not => write!(f, "logical not operator (`!`)"),
+
+            Self::SingleLineComment(_) => write!(f, "single line comment"),
+            Self::MultiLineComment(_) => write!(f, "multi line comment"),
 
             Self::Eof => write!(f, "end-of-file (EOF)"),
         }
@@ -168,10 +204,37 @@ fn float(lex: &mut Lexer<Token>) -> f64 {
     lex.slice().parse().unwrap()
 }
 
+fn single_line_comment(lex: &mut Lexer<Token>) -> String {
+    let remainder = lex.remainder();
+    if let Some(newline_index) = remainder.find('\n') {
+        lex.bump(newline_index + 1);
+        String::from(&remainder[..newline_index])
+    } else {
+        lex.bump(remainder.bytes().len());
+        String::from(remainder)
+    }
+}
+
+fn multi_line_comment(lex: &mut Lexer<Token>) -> Result<String, LexingError> {
+    let remainder = lex.remainder();
+    if let Some(comment_closing_index) = remainder.find("*/") {
+        lex.bump(comment_closing_index + 2);
+        Ok(String::from(&remainder[..comment_closing_index]))
+    } else {
+        let source_len = lex.source().len();
+        Err(LexingError::MissingCommentClosingTag {
+            span: source_len..source_len,
+        })
+    }
+}
+
 #[derive(Clone, Debug, Default, PartialEq)]
 pub enum LexingError {
     IdentifierStartsWithNumber {
         token: String,
+        span: Span,
+    },
+    MissingCommentClosingTag {
         span: Span,
     },
     UnknownToken {
@@ -187,6 +250,7 @@ impl LexingError {
     pub fn get_span(&self) -> Option<Span> {
         match self {
             Self::IdentifierStartsWithNumber { span, .. } => Some(span.clone()),
+            Self::MissingCommentClosingTag { span } => Some(span.clone()),
             Self::UnknownToken { span, .. } => Some(span.clone()),
             _ => None,
         }
@@ -202,6 +266,9 @@ impl Display for LexingError {
                 Self::IdentifierStartsWithNumber { token, .. } => {
                     format!("identifier can't start with number: `{token}`")
                 }
+                Self::MissingCommentClosingTag { .. } => {
+                    "missing comment closing tag (`*/`)".to_string()
+                }
                 Self::UnknownToken { token, .. } => {
                     format!("unknown token `{token}`")
                 }
@@ -214,6 +281,7 @@ impl Display for LexingError {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use indoc::indoc;
 
     #[test]
     fn test_lexing_identifier() {
@@ -271,6 +339,90 @@ mod tests {
 
             assert!(tokens.len() == 2);
             assert_eq!(tokens[0].kind, Token::Float(c.parse().unwrap()));
+        }
+    }
+
+    #[test]
+    fn test_lexing_single_line_comment() {
+        let cases = [
+            indoc! {"
+                // This is a single line comment
+                var x = 5;
+            "},
+            indoc! {"
+                var x = 10;
+
+                print(x); // this should works too!
+            "},
+            indoc! {"
+                // print(y);
+            "},
+            indoc! {"
+            // A single line comment that spans to EOF"},
+        ];
+
+        for c in cases {
+            let lex_result = lex(c);
+
+            assert!(lex_result.is_ok());
+
+            let tokens = lex_result.unwrap();
+
+            assert!(!tokens
+                .iter()
+                .any(|t| matches!(t.kind, Token::SingleLineComment(_))))
+        }
+    }
+
+    #[test]
+    fn test_lexing_multi_line_comment() {
+        let cases = [
+            indoc! {"
+                /**
+                 * This is a multi line comment
+                 */
+                var x = 5;
+            "},
+            indoc! {"
+                var x = 5; /* no problem */
+            "},
+            indoc! {"
+                /**
+                 * No problem too
+                 */
+
+                /**
+                 * print(y);
+                 *       ^
+                 *      A non existing variable
+                 */
+            "},
+        ];
+
+        for c in cases {
+            let lex_result = lex(c);
+
+            assert!(lex_result.is_ok());
+
+            let tokens = lex_result.unwrap();
+
+            assert!(!tokens
+                .iter()
+                .any(|t| matches!(t.kind, Token::MultiLineComment(_))))
+        }
+    }
+
+    #[test]
+    fn test_lexing_invalid_comment() {
+        let cases = [indoc! {"
+                /* No closing tag
+                print(0);
+            "}];
+
+        for c in cases {
+            let lex_result = lex(c);
+
+            assert!(lex_result.is_err());
         }
     }
 }