+++ CHANGELOG.md
@@ -7,9 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Support for more boolean algebra operators.
-- Support for loop.
+- Add assignment operators.
+- Change `print` function behaviour.
 - Change auto-casting behaviour in number types.
+
+## [0.2.2] - 2024-01-11
+
+### Added
+
+- Support for comments.
+- Support for logical boolean operators.
+- Support for loop.
+- Support for modulo (`%`) operator.
 
 ## [0.2.1] - 2024-01-08